Refactor shared library manager to be thread-safe and reorganize code
